move vf rollouts code to more sensible location
